else if문 early return으로 변경 및 외부 함수 파일로 분리
